# openapi.yaml
openapi: 3.0.0
paths:
  /api/v1/health:
    get:
      operationId: HealthController_check_v1
      parameters: []
      responses:
        "200":
          description: The Health Check is successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  info:
                    type: object
                    example: &a1
                      database: &a2
                        status: up
                    additionalProperties:
                      type: object
                      required:
                        - status
                      properties:
                        status:
                          type: string
                      additionalProperties: true
                    nullable: true
                  error:
                    type: object
                    example: {}
                    additionalProperties:
                      type: object
                      required:
                        - status
                      properties:
                        status:
                          type: string
                      additionalProperties: true
                    nullable: true
                  details:
                    type: object
                    example: *a1
                    additionalProperties:
                      type: object
                      required:
                        - status
                      properties:
                        status:
                          type: string
                      additionalProperties: true
        "503":
          description: The Health Check is not successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  info:
                    type: object
                    example: *a1
                    additionalProperties:
                      type: object
                      required:
                        - status
                      properties:
                        status:
                          type: string
                      additionalProperties: true
                    nullable: true
                  error:
                    type: object
                    example:
                      redis: &a3
                        status: down
                        message: Could not connect
                    additionalProperties:
                      type: object
                      required:
                        - status
                      properties:
                        status:
                          type: string
                      additionalProperties: true
                    nullable: true
                  details:
                    type: object
                    example:
                      database: *a2
                      redis: *a3
                    additionalProperties:
                      type: object
                      required:
                        - status
                      properties:
                        status:
                          type: string
                      additionalProperties: true
      tags:
        - Health
  /api/v1/auth/login:
    post:
      operationId: AuthController_login_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "201":
          description: ""
      summary: Log in and receive access and refresh tokens
      tags: &a4
        - auth
  /api/v1/auth/refresh:
    post:
      operationId: AuthController_refresh_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshDto"
      responses:
        "201":
          description: ""
      summary: Rotate tokens using a valid refresh token
      tags: *a4
  /api/v1/auth/password:
    patch:
      operationId: AuthController_changePassword_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordDto"
      responses:
        "200":
          description: ""
      security:
        - bearer: []
      summary: Change the current user password
      tags: *a4
  /api/v1/auth/logout:
    post:
      operationId: AuthController_logout_v1
      parameters: []
      responses:
        "201":
          description: ""
      security:
        - bearer: []
      summary: Log out and revoke the refresh token
      tags: *a4
  /api/v1/publishers:
    get:
      operationId: PublishersController_findAll_v1
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: sortDirection
          required: false
          in: query
          schema:
            default: DESC
            type: string
            enum:
              - ASC
              - DESC
        - name: searchField
          required: false
          in: query
          schema:
            type: string
            enum:
              - name
              - businessNumber
        - name: search
          required: false
          in: query
          description: Search keyword (used with searchField)
          schema:
            type: string
        - name: sortBy
          required: false
          in: query
          schema:
            default: createdAt
            type: string
            enum:
              - name
              - createdAt
      responses:
        "200":
          description: ""
      security: &a5
        - bearer: []
      summary: List publishers (paginated, searchable by name)
      tags: &a6
        - publishers
    post:
      operationId: PublishersController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePublisherDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                example:
                  id: "1"
      security: *a5
      summary: Create a publisher (SUPER_ADMIN only)
      tags: *a6
  /api/v1/publishers/{id}:
    get:
      operationId: PublishersController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a5
      summary: Get a publisher by id
      tags: *a6
    patch:
      operationId: PublishersController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePublisherDto"
      responses:
        "200":
          description: ""
      security: *a5
      summary: Update a publisher (SUPER_ADMIN only)
      tags: *a6
    delete:
      operationId: PublishersController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a5
      summary: Delete a publisher (SUPER_ADMIN only)
      tags: *a6
info:
  title: Bookstore Inventory API
  description: REST API for bookstore chain inventory and replenishment
  version: "1.0"
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    LoginDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: password123
      required:
        - email
        - password
    RefreshDto:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    ChangePasswordDto:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          example: Password123
      required:
        - currentPassword
        - newPassword
    CreatePublisherDto:
      type: object
      properties:
        name:
          type: string
        businessNumber:
          type: string
          example: "1234567890"
        phone:
          type: string
          example: "01012345678"
        email:
          type: string
          example: example@example.com
        zipcode:
          type: string
          example: "06234"
        address:
          type: string
        addressDetail:
          type: string
        contractStartedAt:
          type: string
          example: 2026-01-01
        contractEndedAt:
          type: string
          example: 2027-01-01
      required:
        - name
        - businessNumber
    UpdatePublisherDto:
      type: object
      properties:
        name:
          type: string
        businessNumber:
          type: string
          example: "1234567890"
        phone:
          type: string
          example: "01012345678"
        email:
          type: string
          example: example@example.com
        zipcode:
          type: string
          example: "06234"
        address:
          type: string
        addressDetail:
          type: string
        contractStartedAt:
          type: string
          example: 2026-01-01
        contractEndedAt:
          type: string
          example: 2027-01-01